Made server abstract
